Make protocol.disconnect synchronous Refactor protocol message processing into MessageProcessor (pyotgw/messageprocessor.py) Add and update tests accordingly
